put time series backbones in separate files, add doc strings and refactored search space arguments
